restore shim 322+
